unit tests completed

+++ Challenges/arrayShift/arrayShift/Program.cs
@@ -6,9 +6,9 @@ namespace arrayShift
     {
         public static void Main()
         {
-            decimal[] array = { 2, 4, 6, 8, 9 };
-            decimal middleNum = 5;
-            decimal[] result = ArrayShift(array, middleNum);
+            int[] array = { 2, 4, 6, 8, 9 };
+            int middleNum = 5;
+            int[] result = ArrayShift(array, middleNum);
 
             string arrayAsString = "";
             for (int i = 0; i < array.Length; i++)
@@ -31,11 +31,11 @@ namespace arrayShift
 
         }
 
-        public static decimal[] ArrayShift(decimal[] array, decimal middleNum)
+        public static int[] ArrayShift(int[] array, int middleNum)
         {
-            decimal[] newArray = new decimal[array.Length + 1];
+            int[] newArray = new int[array.Length + 1];
             int counter = 0;
-            decimal midIndex = (decimal) Math.Ceiling((decimal)(array.Length) / 2);
+            decimal midIndex = Math.Ceiling((decimal)(array.Length) / 2);
             for (int i = 0; i < newArray.Length; i++)
             {
                 if (midIndex == i)
